shell: refactor shell_command.py to new util_rules subpackage
