+++ src/python/pants/backend/shell/goals/test.py
@@ -6,14 +6,14 @@ from __future__ import annotations
 import dataclasses
 from typing import Any
 
-from pants.backend.shell import shell_command
-from pants.backend.shell.shell_command import ShellCommandProcessRequest
 from pants.backend.shell.subsystems.shell_test_subsys import ShellTestSubsystem
 from pants.backend.shell.target_types import (
     ShellCommandCommandField,
     ShellCommandTestDependenciesField,
     SkipShellCommandTestsField,
 )
+from pants.backend.shell.util_rules import shell_command
+from pants.backend.shell.util_rules.shell_command import ShellCommandProcessRequest
 from pants.core.goals.test import (
     TestDebugAdapterRequest,
     TestDebugRequest,
