+++ src/python/pants/backend/shell/register.py
@@ -1,7 +1,7 @@
 # Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.backend.shell import dependency_inference, shell_command, shunit2, shunit2_test_runner
+from pants.backend.shell import dependency_inference, shunit2, shunit2_test_runner
 from pants.backend.shell.goals import tailor, test
 from pants.backend.shell.target_types import (
     ShellCommandRunTarget,
@@ -13,6 +13,7 @@ from pants.backend.shell.target_types import (
     Shunit2TestTarget,
 )
 from pants.backend.shell.target_types import rules as target_types_rules
+from pants.backend.shell.util_rules import shell_command
 
 
 def target_types():
